Fix for linux atomics. Fix for ScopedPointer in some compilers. Mac PPC build fix.

+++ src/native/windows/juce_win32_WASAPI.cpp
@@ -451,7 +451,7 @@ public:
                     bufferSize -= samplesToDo;
                     offset += samplesToDo;
 
-                    if (samplesToDo < numSamplesAvailable)
+                    if (samplesToDo < (int) numSamplesAvailable)
                     {
                         reservoirSize = jmin ((int) (numSamplesAvailable - samplesToDo), reservoirCapacity);
                         memcpy ((uint8*) reservoir.getData(), inputData + bytesPerSample * actualNumChannels * samplesToDo,
